requirements/base.txt
SQLAlchemy==1.3.15
pg8000==1.15.2
pyyaml==5.3.1